# openapi-main.yaml
openapi: 3.0.0
servers: 
  - url: 'http://localhost:3000' 
info:
  title: OpenAPI Specification
  version: "1"
...
paths:
...
  /armors/stats:
    post:
      tags:
        - Armors stats
      operationId: createArmorStats
      summary: "Create a new armor stats"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                armor_id:
                  type: integer
                stat_id:
                  type: integer
                value:
                  type: integer
      responses:
        '201':
          description: Armor stats created
        '400':
          description: Invalid input
    get:
      tags:
        - Armors stats
      operationId: listArmorStats
      summary: "List all armor stats"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of armor stats
  /armors/{id}/stats:
    get:
      tags:
        - Armors stats
      operationId: getArmorStats
      summary: "Get armor stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor not found"
    post:
      tags:
        - Armors stats
      summary: "Create a new armor stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer          
      responses:
        '201':
          description: Armor stats created
        '400':
          description: Invalid input
  /armors/{id}/stats/{stat_id}:
    get:
      tags:
        - Armors stats
      operationId: getArmorStatsById
      summary: "Get armor stats by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of armor stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor stats not found"
    put:
      tags:
        - Armors stats
      operationId: updateArmorStats
      summary: "Update an existing armor stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of armor stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor stats not found"
    delete:
      tags:
        - Armors stats
      operationId: deleteArmorStats
      summary: "Delete a armor stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of armor stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor stats not found"
...
  /characters:
    post:
      tags:
        - Characters
      operationId: createCharacter
      summary: "Create a new character"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                user_id:
                  type: integer
                weapon1Id:
                  type: integer
                armor1Id:
                  type: integer
                armor2Id:
                  type: integer
                accessory1Id:
                  type: integer
                accessory2Id:
                  type: integer
      responses:
        '201':
          description: Character created
        '400':
          description: Invalid input
    get:
      tags:
        - Characters
      operationId: listCharacters
      summary: "List all characters"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of characters
  /characters/{id}:
    get:
      tags:
        - Characters
      operationId: getCharacterById
      summary: "Get character by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of character to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
                  user_id:
                    type: integer
                  weapon1Id:
                    type: integer
                  armor1Id:
                    type: integer
                  armor2Id:
                    type: integer
                  accessory1Id:
                    type: integer
                  accessory2Id:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Character not found"
    put:
      tags:
        - Characters
      operationId: updateCharacter
      summary: "Update an existing character"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of character to update"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                user_id:
                  type: integer
                weapon1Id:
                  type: integer
                armor1Id:
                  type: integer
                armor2Id:
                  type: integer
                accessory1Id:
                  type: integer
                accessory2Id:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Character not found"
    delete:
      tags:
        - Characters
      operationId: deleteCharacter
      summary: "Delete a character"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of character to delete"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Character not found"
  /characters/{id}/stats:
    get:
      tags:
        - Characters stats
      operationId: getCharacterStats
      summary: "Get character stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of character to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Character not found"
...
  /characters/{id}/hitpool:
    get:
      tags:
        - Characters stats
      operationId: getCharacterHitpool
      summary: "Get character hitpool"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of character to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  melee_effectiveHitPool:
                    type: integer
                  melee_bosseffectiveHit:
                    type: integer
                  melee_mobeffectiveHit:
                    type: integer
                  ranged_effectiveHitPool:
                    type: integer
                  ranged_bosseffectiveHit:
                    type: integer
                  ranged_mobeffectiveHit:
                    type: integer
                  magic_effectiveHitPool:
                    type: integer
                  magic_bosseffectiveHit:
                    type: integer
                  magic_mobeffectiveHit:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Character not found"
...
  /stats:
    post:
      tags:
        - Stats
      operationId: createStats
      summary: "Create a new stats"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Stats created
        '400':
          description: Invalid input
    get:
      tags:
        - Stats
      operationId: getStats
      summary: "Get stats"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Stats not found"
...